chore(test): refactor settings constants into modular structure
Split monolithic settings.ts into separate files for better organization:
- preferences.ts: all preference-related constants
- registries.ts: all registry-related constants

Export both classes directly from main index.ts.
Update imports to use direct per-class imports.
